# uddr-console/backend/swagger.json
{
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Endpoint\n\nThe base endpoint is `https://api.ddr.ultradns.com`, which all of the following REST routes will follow.\n\n\n# Authentication\n\nAuthentication is performed using X-API-Key header with provided PSK API key on each request\n\n# cURL Example\n\n```curl --header \"X-API-Key: <your API key>\" https://api.ddr.ultradns.com```\n",
    "title": "UltraDDR API Documentation",
    "contact": {},
    "version": "1.0.0"
  },
  "host": "api.ddr.ultradns.com",
  "paths": {
...
    "/userdata/policy/v2/list": {
      "post": {
        "security": [
          {
            "Authorization-Key": [
              "global"
            ]
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "tags": [
          "Policy API"
        ],
        "summary": "list",
        "operationId": "listUsingPOST_3",
        "parameters": [
          {
            "description": "request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ListPoliciesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Policy"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ListPoliciesRequest": {
      "type": "object",
      "title": "ListPoliciesRequest",
      "properties": {
        "organization_id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "Policy": {
      "type": "object",
      "title": "Policy",
      "properties": {
        "created_by": {
          "type": "string"
        },
        "created_on": {
          "type": "string",
          "format": "date-time"
        },
        "deleted_by": {
          "type": "string"
        },
        "deleted_on": {
          "type": "string",
          "format": "date-time"
        },
        "enabled": {
          "type": "boolean"
        },
        "groups": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "modified_by": {
          "type": "string"
        },
        "modified_on": {
          "type": "string",
          "format": "date-time"
        },
        "name": {
          "type": "string"
        },
        "organization_id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Authorization-Key": {
      "type": "apiKey",
      "name": "X-Api-Key",
      "in": "header"
    }
  },
...
}